fix bench

+++ src/timeman.hpp
@@ -9,7 +9,7 @@ constexpr uint64_t INF_TIME = 1 << 28;
 
 struct TimeManager
 {
-    TimeManager(int time = INF_TIME, int inc = 0) {
+    TimeManager(uint64_t time = INF_TIME, uint64_t inc = 0) {
         // convert into microseconds
         time *= 1000;
         inc *= 1000;
